+++ tests/core/calculators/vasp/test_vasp.py
@@ -16,7 +16,7 @@ from pymatgen.io.vasp.sets import MPRelaxSet, MPScanRelaxSet
 
 from quacc import SETTINGS, change_settings
 from quacc.calculators.vasp import Vasp, presets
-from quacc.calculators.vasp.params import MPtoASEConverter
+from quacc.calculators.vasp.params import MPtoASEParams
 from quacc.schemas.prep import prep_next_run
 
 FILE_DIR = Path(__file__).parent
@@ -73,7 +73,7 @@ def test_presets():
     assert calc.exp_params["ediff"] == 1e-5
     assert calc.float_params["encut"] == 450
 
-    parameters = MPtoASEConverter(atoms=atoms).convert_dict_set(MPScanRelaxSet)
+    parameters = MPtoASEParams(atoms=atoms).convert_dict_set(MPScanRelaxSet)
     calc = Vasp(atoms, xc="scan", **parameters)
     assert calc.xc.lower() == "scan"
     assert calc.string_params["algo"].lower() == "all"
@@ -191,7 +191,7 @@ def test_magmoms(atoms_mag, atoms_nomag, atoms_nospin):
 
     atoms = bulk("Cu") * (2, 2, 1)
     atoms[-1].symbol = "Fe"
-    parameters = MPtoASEConverter(atoms=atoms).convert_dict_set(MPScanRelaxSet)
+    parameters = MPtoASEParams(atoms=atoms).convert_dict_set(MPScanRelaxSet)
     calc = Vasp(atoms, **parameters)
     atoms.calc = calc
     assert atoms.get_chemical_symbols() == ["Cu", "Cu", "Cu", "Fe"]
@@ -200,7 +200,7 @@ def test_magmoms(atoms_mag, atoms_nomag, atoms_nospin):
     atoms = bulk("Cu") * (2, 2, 1)
     atoms[-1].symbol = "Fe"
     atoms.set_initial_magnetic_moments([3.14] * (len(atoms) - 1) + [1.0])
-    parameters = MPtoASEConverter(atoms=atoms).convert_dict_set(MPScanRelaxSet)
+    parameters = MPtoASEParams(atoms=atoms).convert_dict_set(MPScanRelaxSet)
     calc = Vasp(atoms, **parameters)
     atoms.calc = calc
     assert atoms.get_initial_magnetic_moments().tolist() == [3.14] * (
@@ -698,7 +698,7 @@ def test_setups():
     assert calc.parameters["setups"]["Cu"] == ""
 
     atoms = bulk("Cu")
-    parameters = MPtoASEConverter(atoms=atoms).convert_dict_set(MPScanRelaxSet)
+    parameters = MPtoASEParams(atoms=atoms).convert_dict_set(MPScanRelaxSet)
     calc = Vasp(atoms, **parameters)
     assert calc.parameters["setups"]["Cu"] == "_pv"
 
@@ -818,12 +818,12 @@ def test_preset_override():
 
 def test_bad_pmg_converter():
     with pytest.raises(ValueError, match="Either atoms or prev_dir must be provided"):
-        MPtoASEConverter()
+        MPtoASEParams()
 
 
 def test_pmg_input_set():
     atoms = bulk("Cu")
-    parameters = MPtoASEConverter(atoms=atoms).convert_dict_set(MPRelaxSet)
+    parameters = MPtoASEParams(atoms=atoms).convert_dict_set(MPRelaxSet)
     calc = Vasp(atoms, incar_copilot="off", **parameters)
     assert calc.parameters == {
         "algo": "fast",
@@ -853,7 +853,7 @@ def test_pmg_input_set():
 def test_pmg_input_set2():
     atoms = bulk("Fe") * (2, 1, 1)
     atoms[0].symbol = "O"
-    parameters = MPtoASEConverter(atoms=atoms).convert_dict_set(MPRelaxSet)
+    parameters = MPtoASEParams(atoms=atoms).convert_dict_set(MPRelaxSet)
     calc = Vasp(atoms, incar_copilot="off", **parameters)
     assert calc.parameters == {
         "algo": "fast",